Many copy-button changes
Now has a svg-file as the icon for the button. Created by creating a
svgNode that we fill with a svg-path. Which is then added to our button.

Fix for copy-button not copying correct text.
Now find text by using querySelector(#text-card-paragraph).textContent.

Fix for "cannot read property of controls"-error, by providing "mod"
into the scroll into the renderTextCards-function.