Fix for marking
If something is already marked, and the user marks something else, this
becomes a new marking, instead of being added to the marking.

User can now ctrl + click to add to the current markin,
just like in the native visualizations on spotfire.
If ctrl + click on an already selected card, it is removed,
just like the native ones.