Fix scroll won't activate when scroll
Changed a bunch of stuff from onmousedown to onmouseup.
Got the X-position of mouse to not react on marking if in the same
position of scroll.

New Bug: Can't mark when trying to mark on the position where a
scrollbar should be, even though there are no scroll bar there.